requirements.txt
torch
numpy
opencv-python
mtcnn
mediapipe
tqdm
PyYAML
scikit-learn 